Move DTOs to the parent folder/namespace
